src/app/pdf-viewer/pdf-viewer.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { PdfViewerComponent } from './pdf-viewer.component';

describe('PdfViewerComponent', () => {
  let component: PdfViewerComponent;
  let fixture: ComponentFixture<PdfViewerComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ PdfViewerComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(PdfViewerComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it(`should have as title 'issue'`, () => {
    const fixture = TestBed.createComponent(PdfViewerComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('issue');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(PdfViewerComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.content span').textContent).toContain('issue226 app is running!');
  });
});
